working on profile landing page
